Basic cached building, with diagnostics
Some rule files fail parsing.

Going through a, b, ... with:

    cd metadata ; zl=(b*); zones=${(j:,:)zl%.json}; cd ..; go run build/cmd/zonedb/main.go -v -update-idn -zones $zones -w

we fail in zones starting `a` or `b` on:

    aquarelle_latn_1.txt
    bzh_fr_1.txt
    bostik_latn_1.txt

In short: we assume lines with comments so that we can look past the end
of a line.  That's wrong.  Next commit includes the metadata so far, the
one after will fix this bug.